+++ AppLovin/CHANGELOG.md
@@ -1,7 +1,8 @@
 ## Changelog
    * 6.1.4.2
-     * Addressed various issues introduced in the last version:
+     * Addressed some issues and optimized adapter:
          * Allow passing in of SDK key via the initialization `configuration` object as many publishers integrate without SDK key in the `Info.plist`.
+         * Add support for using a cached `configuration` NSDictionary for initialization.
          * Do not consider banner ad display failure as ad load failure.
          * Do not consider users closing rewarded ad early or reward validation failure as an ad load failure.